+++ api/docs/CHANGELOG.md
@@ -21,6 +21,9 @@ keywords: "API, Docker, rcli, REST, documentation"
   when negotiating content-type.
 * Deprecated: The `POST /grpc` and `POST /session` endpoints are deprecated and
   will be removed in a future version.
+* `POST /containers/create` accepts a new `apisocket` Mount type to automatically expose engine socket inside container.
+  `access` let user select the mechanism used to expose the engine socket to container. `unconfined` mounts the engine
+  socket without any restriction.
 
 ## v1.52 API changes